Changed the segue to show the info post view modally
Declared the textfield delegate to dismiss keyboard on return

Added a method to dismiss the keyboard by tapping on screen

Tried to implement reachability, but the simulator does not change the
error message. I had read that it might be a problem in simulator, but
would like that code reviewed.

Used updateAt constant to order results by most recent

Added the map delegate to the Map View Controller to allow student URLs
to be linked